修改项目收款单Bug（阳）

+++ hzaction-master/src/main/java/com/action/actaccount/controller/VwBusCollectionController.java
@@ -71,7 +71,7 @@ import org.jeecgframework.web.cgform.service.config.CgFormFieldServiceI;
 import java.util.HashMap;
 /**   
  * @Title: Controller  
- * @Description: 项目收款单视图
+ * @Description: 项目收款单
  * @author onlineGenerator
  * @date 2019-08-22 15:34:59
  * @version V1.0   
@@ -98,7 +98,7 @@ public class VwBusCollectionController extends BaseController {
 
 	@Autowired
 	/**
-	 * 项目收款单视图列表 页面跳转
+	 * 项目收款单列表 页面跳转
 	 * 
 	 * @return
 	 */
@@ -141,7 +141,7 @@ public class VwBusCollectionController extends BaseController {
 	
 	
 	/**
-	 * 删除项目收款单视图
+	 * 删除项目收款单
 	 * 
 	 * @return
 	 */
@@ -150,22 +150,25 @@ public class VwBusCollectionController extends BaseController {
 	public AjaxJson doDel(VwBusCollectionEntity vwBusCollection, HttpServletRequest request) {
 		String message = null;
 		AjaxJson j = new AjaxJson();
-		vwBusCollection = systemService.getEntity(VwBusCollectionEntity.class, vwBusCollection.getId());
-		message = "项目收款单视图删除成功";
+//		vwBusCollection = systemService.getEntity(VwBusCollectionEntity.class, vwBusCollection.getId());
+		message = "项目收款单删除成功";
 		try{
-			BusCollectionEntity busConEntity = new 	BusCollectionEntity();
-			MyBeanUtils.copyBeanNotNull2Bean(vwBusCollection ,busConEntity);
+			BusCollectionEntity busConEntity = systemService.getEntity(BusCollectionEntity.class, vwBusCollection.getId());
+//			MyBeanUtils.copyBeanNotNull2Bean(vwBusCollection ,busConEntity);
 			BusCollectionService.delete(busConEntity);
 			// 删除支付
-			BusPayInfoEntity busPayInfoEntity = new BusPayInfoEntity();
-			MyBeanUtils.copyBeanNotNull2Bean(vwBusCollection ,busPayInfoEntity);
-			busPayInfoEntity.setId(vwBusCollection.getFromPayinfoId());
-			busPayInfoService.delete(busPayInfoEntity);
+			if(StringUtil.isNotEmpty(vwBusCollection.getFromPayinfoId())) {
+				BusPayInfoEntity busPayInfoEntity = systemService.getEntity(BusPayInfoEntity.class, vwBusCollection.getFromPayinfoId());
+//			MyBeanUtils.copyBeanNotNull2Bean(vwBusCollection ,busPayInfoEntity);
+//			busPayInfoEntity.setId(vwBusCollection.getFromPayinfoId());
+				busPayInfoService.delete(busPayInfoEntity);
+			}
+	
 			
 			systemService.addLog(message, Globals.Log_Type_DEL, Globals.Log_Leavel_INFO);
 		}catch(Exception e){
 			e.printStackTrace();
-			message = "项目收款单视图删除失败";
+			message = "项目收款单删除失败";
 			throw new BusinessException(e.getMessage());
 		}
 		j.setMsg(message);
@@ -173,7 +176,7 @@ public class VwBusCollectionController extends BaseController {
 	}
 	
 	/**
-	 * 批量删除项目收款单视图
+	 * 批量删除项目收款单
 	 * 
 	 * @return
 	 */
@@ -182,7 +185,7 @@ public class VwBusCollectionController extends BaseController {
 	public AjaxJson doBatchDel(String ids,HttpServletRequest request){
 		String message = null;
 		AjaxJson j = new AjaxJson();
-		message = "项目收款单视图删除成功";
+		message = "项目收款单删除成功";
 		try{
 			for(String id:ids.split(",")){
 
@@ -197,7 +200,7 @@ public class VwBusCollectionController extends BaseController {
 			}
 		}catch(Exception e){
 			e.printStackTrace();
-			message = "项目收款单视图删除失败";
+			message = "项目收款单删除失败";
 			throw new BusinessException(e.getMessage());
 		}
 		j.setMsg(message);
@@ -206,7 +209,7 @@ public class VwBusCollectionController extends BaseController {
 
 
 	/**
-	 * 添加项目收款单视图
+	 * 添加项目收款单
 	 * 
 	 * @param ids
 	 * @return
@@ -220,7 +223,7 @@ public class VwBusCollectionController extends BaseController {
 
 //		List<>
 //		List<VmMergeBusPoApplyDetailEntity>
-		message = "项目收款单视图添加成功";
+		message = "项目收款单添加成功";
 		BusCollectionEntity busConEntity = new 	BusCollectionEntity();
 		try{
 			MyBeanUtils.copyBeanNotNull2Bean(vwBusCollection ,busConEntity);
@@ -230,7 +233,7 @@ public class VwBusCollectionController extends BaseController {
 			systemService.addLog(message, Globals.Log_Type_INSERT, Globals.Log_Leavel_INFO);
 		}catch(Exception e){
 			e.printStackTrace();
-			message = "项目收款单视图添加失败";
+			message = "项目收款单添加失败";
 			throw new BusinessException(e.getMessage());
 		}
 		j.setMsg(message);
@@ -239,7 +242,7 @@ public class VwBusCollectionController extends BaseController {
 	}
 	
 	/**
-	 * 更新项目收款单视图
+	 * 更新项目收款单
 	 * 
 	 * @param ids
 	 * @return
@@ -249,7 +252,7 @@ public class VwBusCollectionController extends BaseController {
 	public AjaxJson doUpdate(VwBusCollectionEntity vwBusCollection, HttpServletRequest request) {
 		String message = null;
 		AjaxJson j = new AjaxJson();
-		message = "项目收款单视图更新成功";
+		message = "项目收款单更新成功";
 //		VwBusCollectionEntity t = vwBusCollectionService.findUniqueByProperty(VwBusCollectionEntity.class, "id",vwBusCollection.getId());
 		
 		
@@ -289,7 +292,7 @@ public class VwBusCollectionController extends BaseController {
 			systemService.addLog(message, Globals.Log_Type_UPDATE, Globals.Log_Leavel_INFO);
 		} catch (Exception e) {
 			e.printStackTrace();
-			message = "项目收款单视图更新失败";
+			message = "项目收款单更新失败";
 			throw new BusinessException(e.getMessage());
 		}
 		j.setMsg(message);
@@ -299,7 +302,7 @@ public class VwBusCollectionController extends BaseController {
 	
 
 	/**
-	 * 项目收款单视图新增页面跳转
+	 * 项目收款单新增页面跳转
 	 * 
 	 * @return
 	 */
@@ -313,7 +316,7 @@ public class VwBusCollectionController extends BaseController {
 		return new ModelAndView("com/action/actaccount/vwBusCollection-add");
 	}
 	/**
-	 * 项目收款单视图编辑页面跳转
+	 * 项目收款单编辑页面跳转
 	 * 
 	 * @return
 	 */
@@ -378,9 +381,9 @@ public class VwBusCollectionController extends BaseController {
 		CriteriaQuery cq = new CriteriaQuery(VwBusCollectionEntity.class, dataGrid);
 		org.jeecgframework.core.extend.hqlsearch.HqlGenerateUtil.installHql(cq, vwBusCollection, request.getParameterMap());
 		List<VwBusCollectionEntity> vwBusCollections = this.vwBusCollectionService.getListByCriteriaQuery(cq,false);
-		modelMap.put(NormalExcelConstants.FILE_NAME,"项目收款单视图");
+		modelMap.put(NormalExcelConstants.FILE_NAME,"项目收款单");
 		modelMap.put(NormalExcelConstants.CLASS,VwBusCollectionEntity.class);
-		modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("项目收款单视图列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
+		modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("项目收款单列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
 			"导出信息"));
 		modelMap.put(NormalExcelConstants.DATA_LIST,vwBusCollections);
 		return NormalExcelConstants.JEECG_EXCEL_VIEW;
@@ -394,9 +397,9 @@ public class VwBusCollectionController extends BaseController {
 	@RequestMapping(params = "exportXlsByT")
 	public String exportXlsByT(VwBusCollectionEntity vwBusCollection,HttpServletRequest request,HttpServletResponse response
 			, DataGrid dataGrid,ModelMap modelMap) {
-    	modelMap.put(NormalExcelConstants.FILE_NAME,"项目收款单视图");
+    	modelMap.put(NormalExcelConstants.FILE_NAME,"项目收款单");
     	modelMap.put(NormalExcelConstants.CLASS,VwBusCollectionEntity.class);
-    	modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("项目收款单视图列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
+    	modelMap.put(NormalExcelConstants.PARAMS,new ExportParams("项目收款单列表", "导出人:"+ResourceUtil.getSessionUser().getRealName(),
     	"导出信息"));
     	modelMap.put(NormalExcelConstants.DATA_LIST,new ArrayList());
     	return NormalExcelConstants.JEECG_EXCEL_VIEW;
